requirements.txt
python-telegram-bot==13.15
google-generativeai==0.8.3
redis==5.0.1
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1
chardet==5.2.0
requests==2.31.0
